Move id_generator to it's own module
